backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# LangChain 관련
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-openai==0.0.5

# 벡터 데이터베이스
chromadb==0.4.18

# 검색 및 임베딩
sentence-transformers==2.2.2
rank-bm25==0.2.2

# 기타
numpy>=1.26.0
openai==1.3.0 